XIVY-11521 Add more actions
- Create process (sub, trigger)

- Open endpage

- Create useAction hook
