+++ packages/editor/src/components/parts/call/dialog/DialogCallPart.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useMemo, useState } from 'react';
-import { useClient, useEditorContext, useValidation } from '../../../../context';
+import { useAction, useClient, useEditorContext, useValidation } from '../../../../context';
 import { PartProps, usePartDirty, usePartState } from '../../../props';
 import { CallData, CallableStart, DialogCallData, InscriptionValidation, VariableInfo } from '@axonivy/inscription-protocol';
 import CallMapping from '../CallMapping';
@@ -7,7 +7,6 @@ import { useCallData, useDialogCallData } from '../useCallData';
 import CallSelect from '../CallSelect';
 import { IvyIcons } from '@axonivy/editor-icons';
 import { Fieldset, FieldsetControl, useFieldset } from '../../../../components/widgets';
-import { NewHtmlDialogAction } from './new-html-dialog-action';
 
 function useCallPartValidation(): InscriptionValidation[] {
   const dialog = useValidation('config/dialog');
@@ -44,15 +43,12 @@ const DialogCallPart = () => {
     [config.dialog, startItems]
   );
 
+  const action = useAction('newHtmlDialog');
+  const createDialog: FieldsetControl = { label: 'Create new Html Dialog', icon: IvyIcons.Add, action: () => action() };
   const callField = useFieldset();
-
-  const fieldsetControls: FieldsetControl[] = [
-    { label: 'Create new Html Dialog', icon: IvyIcons.Add, action: () => client.action(NewHtmlDialogAction.create(editorContext.pid)) }
-  ];
-
   return (
     <>
-      <Fieldset label='Dialog' message={dialogValidation} {...callField.labelProps} controls={fieldsetControls}>
+      <Fieldset label='Dialog' message={dialogValidation} {...callField.labelProps} controls={[createDialog]}>
         <CallSelect
           start={config.dialog}
           onChange={change => update('dialog', change)}
