--- packages/editor/src/components/parts/call/dialog/new-html-dialog-action.ts
@@ -1,13 +1,0 @@
-import { InscriptionAction, PID } from '@axonivy/inscription-protocol';
-
-export namespace NewHtmlDialogAction {
-  export const KIND = 'newHtmlDialog';
-
-  export function create(pid: PID): InscriptionAction {
-    return {
-      kind: KIND,
-      pid,
-      payload: ''
-    };
-  }
-}
