+++ packages/editor/src/components/widgets/fieldset/fieldset-control.ts
@@ -1,0 +1,8 @@
+import { IvyIcons } from '@axonivy/editor-icons';
+
+export interface FieldsetControl {
+  label: string;
+  icon: IvyIcons;
+  active?: boolean;
+  action: () => void;
+}
