+++ packages/editor/src/components/widgets/fieldset/index.ts
@@ -1,3 +1,4 @@
 export { default as Fieldset } from './Fieldset';
 export { default as SummaryFieldset } from './SummaryFieldset';
-export * from './Fieldset';
+export * from './fieldset-control';
+export * from './useFieldset';
