+++ packages/editor/src/context/index.ts
@@ -3,3 +3,4 @@ export * from './useDataContext';
 export * from './useValidation';
 export * from './useEditorContext';
 export * from './useTheme';
+export * from './useAction';
