+++ packages/protocol/src/actions.ts
@@ -1,0 +1,5 @@
+export interface InscriptionActionHandlers {
+  newHtmlDialog: undefined;
+  newProcess: undefined;
+  openFile: string;
+}
