+++ packages/protocol/src/index.ts
@@ -2,3 +2,4 @@ export * from './inscription-client';
 export * from './inscription-protocol';
 export * from './data';
 export * from './defaults';
+export * from './actions';
